+++ docs/changelog.rst
@@ -14,6 +14,8 @@ Changelog
 - Fix using remote control to set cursor text color causing errors when
   creating new windows (:iss:`1326`)
 
+- Fix window title for minimized windows not being updated (:iss:`1332`)
+
 
 0.13.3 [2019-01-19]
 ------------------------------